Improved TravisCI integration and dependency graph in composer.json
I updated the build scripts to use the composer.json. This way we can remove
the duplicate ways of registering the dependencies. Using composer also means
we can lose the bootstrap and custom autoload configs.

For building in TravisCI i added code coverage. First of all this gives a great
insight in how well we are testing, but it also parses all our php file so it
will fail on missing dependencies, and (untested) syntax errors.

There is only one thing missing at this point, and that is the version numbers
for the Kunstmaan bundles.